backend/requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pymongo==4.7.2
pinecone==4.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
streamlit==1.35.0
PyJWT==2.9.0
requests==2.32.2
transformers==4.46.3
torch==2.5.1
torchvision==0.20.1